Move find_heuristic into its own file so registration works
