+++ app_allocator/classes/allocator.py
@@ -11,7 +11,7 @@ from random import choice
 
 from app_allocator.classes.application import Application
 from app_allocator.classes.assignments import assign
-from app_allocator.classes.heuristic import find_heuristic
+from app_allocator.classes.find_heuristic import find_heuristic
 from app_allocator.classes.judge import Judge
